+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 ## azure-maps-overview-map Changelog
 
+<a name="0.0.3"></a>
+# 0.0.3 (2024-08-26)
+
+- Improvement: Adjusted toggle button position when overview map shape is round.
+- Bug fix: Fixed exception that occured when passing in options when initializing the control.
+- Updated dependancies.
+   
 <a name="0.0.2"></a>
-# 0.0.1 (2021-02-11)
+# 0.0.2 (2021-02-11)
 
 **Features**